webxr/xrSession_requestAnimationFrame_callback_calls.https.html

<!DOCTYPE html>
<meta name="timeout" content="long">
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src="resources/webxr_util.js"></script>
  <canvas></canvas>

  <script>
    let immersiveTestName = "XRSession requestAnimationFrame calls the " +
      "provided callback for an immersive session";
    let nonImmersiveTestName = "XRSession requestAnimationFrame calls the " +
      "provided callback a non-immersive session";

    let fakeDeviceInitParams = { supportsImmersive:true };

    let testFunction = (session) => new Promise((resolve) => {


      function logMsg(message) {
        console.error(session.mode + ": " + message);
      }

      logMsg("start");

      function onFrame(time, xrFrame) {
        logMsg("onFrame");
        if (xrFrame instanceof XRFrame) {
          logMsg("assert valid");
        } else {
          logMsg("assert invalid");
        }
        assert_true(xrFrame instanceof XRFrame);
        // Test does not complete until the returned promise resolves.
        resolve();
      }

      logMsg("requestingAnimationFrame");
      session.requestAnimationFrame(onFrame);
    });

    xr_session_promise_test(immersiveTestName, testFunction,
      fakeDeviceInitParams, 'immersive-vr');
    xr_session_promise_test(nonImmersiveTestName, testFunction,
      fakeDeviceInitParams, 'inline');

  </script>
</body>
