third_party/blink/web_tests/platform/mac-mac15-arm64/external/wpt/webxr/xrSession_requestAnimationFrame_callback_calls.https-expected.txt

This is a testharness.js-based test.
[FAIL] XRSession requestAnimationFrame calls the provided callback for an immersive session - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] XRSession requestAnimationFrame calls the provided callback for an immersive session - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] XRSession requestAnimationFrame calls the provided callback a non-immersive session - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] XRSession requestAnimationFrame calls the provided callback a non-immersive session - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
Harness: the test ran to completion.
